# backend/openapi.json
{
  "components": {
    "schemas": {
      "AttachmentResponse": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_image": {
            "title": "Is Image",
            "type": "boolean"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          },
          "post_id": {
            "title": "Post Id",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          }
        },
        "required": [
          "id",
          "post_id",
          "original_name",
          "content_type",
          "size",
          "is_image"
        ],
        "title": "AttachmentResponse",
        "type": "object"
      },
      "BoardCreateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "read_visibility": {
            "$ref": "#/components/schemas/ReadVisibility"
          },
          "slug": {
            "maxLength": 200,
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9-]*$",
            "title": "Slug",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/BoardType"
          }
        },
        "required": [
          "name",
          "slug",
          "type",
          "read_visibility"
        ],
        "title": "BoardCreateRequest",
        "type": "object"
      },
      "BoardResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "read_visibility": {
            "$ref": "#/components/schemas/ReadVisibility"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/BoardType"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "type",
          "read_visibility",
          "description",
          "created_at"
        ],
        "title": "BoardResponse",
        "type": "object"
      },
      "BoardType": {
        "enum": [
          "NOTICE",
          "GENERAL",
          "IMAGE"
        ],
        "title": "BoardType",
        "type": "string"
      },
      "Body_create_post_with_attachments_boards__board_id__posts_with_attachments_post": {
        "properties": {
          "content": {
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "files": {
            "default": [],
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ],
        "title": "Body_create_post_with_attachments_boards__board_id__posts_with_attachments_post",
        "type": "object"
      },
...
      "ContentStatus": {
        "description": "A-03 / ADR-0005: only COMMITTED rows are served to users.",
        "enum": [
          "PENDING",
          "COMMITTED"
        ],
        "title": "ContentStatus",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "OriginalUrlResponse": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "OriginalUrlResponse",
        "type": "object"
      },
...
      "PostResponse": {
        "properties": {
          "author_id": {
            "title": "Author Id",
            "type": "integer"
          },
          "board_id": {
            "title": "Board Id",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "board_id",
          "author_id",
          "title",
          "content",
          "status",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "PostResponse",
        "type": "object"
      },
...
      "ReadVisibility": {
        "enum": [
          "PUBLIC",
          "AUTHENTICATED"
        ],
        "title": "ReadVisibility",
        "type": "string"
      },
      "Role": {
        "enum": [
          "USER",
          "ADMIN"
        ],
        "title": "Role",
        "type": "string"
      },
      "SignupRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "SignupRequest",
        "type": "object"
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
      "UserResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "created_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Boards API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/boards": {
      "post": {
        "operationId": "create_board_admin_boards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Board",
        "tags": [
          "boards"
        ]
      }
    },
    "/admin/boards/{board_id}": {
      "delete": {
        "operationId": "delete_board_admin_boards__board_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Board",
        "tags": [
          "boards"
        ]
      }
    },
    "/admin/users/{user_id}/promote": {
      "post": {
        "description": "Y-02: an existing ADMIN promotes a USER to ADMIN.",
        "operationId": "promote_user_admin_users__user_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Promote User",
        "tags": [
          "auth"
        ]
      }
    },
    "/attachments/{attachment_id}/original-url": {
      "get": {
        "operationId": "get_original_url_attachments__attachment_id__original_url_get",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "title": "Attachment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OriginalUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Original Url",
        "tags": [
          "attachments"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "signup_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "auth"
        ]
      }
    },
    "/boards": {
      "get": {
        "operationId": "list_boards_boards_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BoardResponse"
                  },
                  "title": "Response List Boards Boards Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Boards",
        "tags": [
          "boards"
        ]
      }
    },
    "/boards/{board_id}": {
      "get": {
        "operationId": "get_board_boards__board_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Board",
        "tags": [
          "boards"
        ]
      }
    },
...
    "/boards/{board_id}/posts/with-attachments": {
      "post": {
        "operationId": "create_post_with_attachments_boards__board_id__posts_with_attachments_post",
        "parameters": [
          {
            "in": "path",
            "name": "board_id",
            "required": true,
            "schema": {
              "title": "Board Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_post_with_attachments_boards__board_id__posts_with_attachments_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Post With Attachments",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/posts/{post_id}/attachments": {
      "get": {
        "operationId": "list_attachments_posts__post_id__attachments_get",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttachmentResponse"
                  },
                  "title": "Response List Attachments Posts  Post Id  Attachments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Attachments",
        "tags": [
          "attachments"
        ]
      }
    },
    "/posts/{post_id}/attachments/{attachment_id}": {
      "delete": {
        "operationId": "delete_attachment_posts__post_id__attachments__attachment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "title": "Attachment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Attachment",
        "tags": [
          "attachments"
        ]
      }
    },
...
  }
}